Use a dedicated download path when downloading a shared deck
This commit declares a new "external-files-path" in the paths declared
for FileProvider and changes the code related to shared decks to
accomodate the new path.

The previous code constructed the FileProvider by using getExternalFilesDir()
but we didn't have a "external-files-path" correspondent in
filepaths.xml so I think it reverted/defaulted to:

<external-path path="." name="photos" />

which may result in a NotFoundException.

See documentation related to FileProvider file paths:
https://developer.android.com/reference/androidx/core/content/FileProvider